Extract an Interface from ConfigClient, moving the implementation into ConfigClientImpl
