+++ src/test/java/cloud/prefab/client/config/logging/Log4j2ConfigListenerTest.java
@@ -2,7 +2,7 @@ package cloud.prefab.client.config.logging;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import cloud.prefab.client.ConfigClient;
+import cloud.prefab.client.ConfigClientImpl;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.core.config.Configurator;
@@ -17,7 +17,10 @@ public class Log4j2ConfigListenerTest extends AbstractLoggingListenerTest {
 
   @Test
   public void itSetsSpecificLogLevel() {
-    new ConfigClient(clientWithSpecificLogLevel(), Log4j2ConfigListener.getInstance());
+    new ConfigClientImpl(
+      clientWithSpecificLogLevel(),
+      Log4j2ConfigListener.getInstance()
+    );
 
     assertThat(LogManager.getLogger(specificLoggerName()).getLevel())
       .isEqualTo(Level.WARN);
@@ -27,7 +30,7 @@ public class Log4j2ConfigListenerTest extends AbstractLoggingListenerTest {
 
   @Test
   public void itSetsDefaultLogLevel() {
-    new ConfigClient(clientWithDefaultLogLevel(), Log4j2ConfigListener.getInstance());
+    new ConfigClientImpl(clientWithDefaultLogLevel(), Log4j2ConfigListener.getInstance());
 
     assertThat(LogManager.getLogger(specificLoggerName()).getLevel())
       .isEqualTo(Level.WARN);
